+++ AdditiveWindSynth/src/voice.h
@@ -9,12 +9,12 @@ struct GlobalOscillatorModifiers
 {
   float attackScale{1.f};
   float releaseScale{1.f};
-  float pitchOffsetCents{0.f};
+  double pitchOffsetCents{0.0};
   float panOffset{0.f};
   float intensityVariationAmplitudeScale{1.f};
   float intensityVariationRateScale{1.f};
-  float pitchVariationAmplitudeScale{1.f};
-  float pitchVariationRateScale{1.f};
+  double pitchVariationAmplitudeScale{1.0};
+  double pitchVariationRateScale{1.0};
   float panVariationAmplitudeScale{1.f};
   float panVariationRateScale{1.f};
   float portamentoTimeAtCC5MinSec{0.f};
@@ -29,9 +29,9 @@ public:
 
   SynthVoice();
   bool IsActive() const;
-  void Start(float pitch, float pitchBend, float breath);
+  void Start(double pitch, double pitchBend, float breath);
   void Stop();
-  void SetPitchBend(float pitchBend);
+  void SetPitchBend(double pitchBend);
   void SetBreath(float breath);
   void SetPortamentoControl(float control);
   void SetGlobalOscillatorModifiers(const GlobalOscillatorModifiers& modifiers);
@@ -43,17 +43,17 @@ public:
 private:
   void UpdatePitch();
   void UpdateLevels();
-  float GetPortamentoTimeSec() const;
+  double GetPortamentoTimeSec() const;
   float SmoothBreath(float breath);
-  void ApplyOscillatorSettings(int harmonic, const OscillatorSettings& settings, float fundamentalPitchCents);
+  void ApplyOscillatorSettings(int harmonic, const OscillatorSettings& settings, double fundamentalPitchCents);
 
   static constexpr int kNumHarmonics = SimplePreset::kNumOscillators;
 
   // Pitch is in MIDI note numbers (0-127), where 69 corresponds to A4 (440 Hz).
-  float mPitch{0.f};
+  double mPitch{0.0};
 
   // Pitch bend is a semitone offset in MIDI-note units.
-  float mPitchBend{0.f};
+  double mPitchBend{0.0};
 
   // Breath is a linear value from 0 to 1.
   float mBreath{0.f};
